Add date filter and profit container
I added a date filter that will select leads in a range. It uses redux
to filter the leads. I added a profit component that will only show if
you have filtered. The functionality is still not implemented but I
guess we can get the sum of values by using reduce on the filtered array
of leads.